helical_cone_curve_3d/requirements.txt
numpy
scipy
matplotlib
pyyaml